Force set device name
Makes it possible to set the device name for a PZP with
--force-device-name=NAME upon startup. This is needed, so it's possible
to start more than one PZP on a single host as the device name will be
based on the hostname by default. The commit changes the various config
methods that are called to pass that custom device name and use it.

It became appearant that sometimes device name and friendly name are mixed
up and can mean the same. This commit changes the code for the PZP to only
use a device name, the friendly name might be only stored in the config
file, but never displayed. For the PZH friendly name is still essentially
the device name.

Also removes some no longer used arguments from the PZP start up script.

Jira-issue: WP-250