Fix enrolment in line with PZH user identity changes
A number of changes at the PZH and PZH web server mean that
the PZP was not enrolling properly.  This has now been fixed.

In addition, the PZP will now report when enrolment was a
success or not (at least in some cases) and the new PZH and PZHWS
code should be a bit friendlier.

Jira issue: WP-959