Prevent webinos.js from loading multiple times
Check if webinos already exists on a page so that it will not reload and reinitialize everything.

This will allow the apps to include a webinos.js so that it can run both inside a renderer, where we inject it, as well as on a regular web browser.

This is also solved by referencing blindly (concerning the port) webinos on http://localhost:8080/webinos.js

Jira issue: WP-489